% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{server}
\alias{server}
\title{Server function for the shiny app.}
\usage{
server(input, output, session)
}
\arguments{
\item{input}{from the ui object}

\item{output}{for the ui object}

\item{session}{for observe object}
}
\description{
Server function for the shiny app.
}
