% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinydistributions.R
\docType{package}
\name{shinydistributions}
\alias{shinydistributions}
\alias{shinydistributions-package}
\title{shinydistributions: A shiny-app to show which distribution fits best.}
\description{
shinydistributions: A shiny-app to show which distribution fits best.
}
